Create neo4j repositories (#30)
* add mongodb context and create local mongodb database script

* add some more sample flight data

* inject repositories into services depening on db type and mongodb-seeder added

* add MongoDB entities

* add FlightMongoRepository and mappings from FlightMongo to Flight

* encountered problem mapping BookingMongo to Booking

* add mappings for BookingMongo, embedded documents and snapshot documents to Booking

* add UserRepository

* change all primary and foreign keys to long instead of int

* forgot to change int to long in the stored procedure code

* register IMongoClient for the MongoDBSeeder and create index on booking.tickets.flight.id

* AirportMongoDBRepository and AirlineMongoDBRepository added

* create triggers on flight insert, update and delete

* create mongo dump script

* the delete trigger now uses correct deletedBy email

* Add Cypher script for airline system and Docker setup script

* change to async transaction start instead of synchronous

* add another cypher script

* add Version to flight to implement optimistic concurrency

* increment version on update

* set testdata script to mongo db

* fix search for flights so that it uses index correctly

* changes

* script to create neo4j docker container

* make some models

* zz

* add some repositories

* create flight repo and all get flight repo

* make some changes in model

* Add alle Neo4j repositories

* fix endpoint url and improve sample Neo4j data

---------

+++ backend/Dtos/BookingCreationRequest.cs
@@ -16,8 +16,8 @@ namespace backend.Dtos
     {
         public PassengerCreationRequest Passenger { get; set; }
 
-        public int FlightId { get; set; }
+        public long FlightId { get; set; }
 
-        public int FlightClassId { get; set; }
+        public long FlightClassId { get; set; }
     }
 }
